Refactoring after PR 16434: rename schemeshard_import_scheme_getter.h -> schemeshard_import_getters.h
